updated sales with unit, stock sold check with unit, unit comversion rate

+++ database/migrations/2019_05_01_185239_create_stocks_table.php
@@ -22,7 +22,7 @@ class CreateStocksTable extends Migration
             $table->double('quantity');
             $table->unsignedInteger('no_of_jar')->default(0);
             $table->unsignedInteger('no_of_drum')->default(0);
-            $table->integer('sold')->default(0);
+            $table->double('sold')->default(0);
             $table->double('price');
             $table->double('sale_price')->default(0);
             $table->tinyInteger('status')->default(1);
